splitting the cluster.go file to extends its status
